Move search component type definitions to a separate file
